+++ src/client/src/pages/home/homeComponent.js
@@ -65,6 +65,12 @@ class HomeComponent extends React.Component {
           moveend={(extent, resolution) => {
             this.props.filterByExtent(extent, resolution);
           }}
+          filterByExtent={extent => {
+            this.props.filterByExtent(extent);
+          }}
+          setmapfilter={checked => {
+            this.props.setmapfilter(checked);
+          }}
           selected={id => {
             if (id === null) {
               history.push(process.env.PUBLIC_URL);
@@ -576,6 +582,12 @@ const mapDispatchToProps = dispatch => {
         type: "SEARCH_EXTENT_CHANGED",
         extent: extent,
         resolution: resolution,
+      });
+    },
+    setmapfilter: active => {
+      dispatch({
+        type: "SEARCH_MAPFILTER_CHANGED",
+        active: active,
       });
     },
     resetCart: () => {